requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.39.0
supabase==2.28.0
pydantic-settings==2.11.0
pyjwt[crypto]==2.11.0
httpx==0.28.1
python-dotenv==1.2.1
python-multipart==0.0.20
